add again missing files to project
