+++ KarhooUISDK/Screens/CheckoutScreen/NewCheckout/ViewModels/TrainNumberCellViewModel.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 class TrainNumberCellViewModel: DetailsCellViewModel {
-    init(onTap: @escaping () -> Void){
+    init(onTap: @escaping () -> Void) {
         super.init(
             title: UITexts.Booking.trainTitle,
             subtitle: UITexts.Booking.trainSubtitle,
