+++ KarhooUISDK/Screens/CheckoutScreen/NewCheckout/ViewModels/VehicleDetailsCardViewModel.swift
@@ -1,0 +1,18 @@
+//
+//  VehicleDetailsCardViewModel.swift
+//  KarhooUISDK
+//
+//  Created by Bartlomiej Sopala on 13/01/2023.
+//  Copyright © 2023 Flit Technologies Ltd. All rights reserved.
+//
+
+import Foundation
+
+struct VehicleDetailsCardViewModel {
+    let title: String
+    let passengerCapacity: Int
+    let luggageCapacity: Int
+    let fleetName: String
+    let carIconUrl: String
+    let fleetIconUrl: String
+}
